+++ Dockerfile
@@ -1,7 +1,10 @@
-## Build
-FROM golang:1.18-alpine AS build
+## Base
+FROM golang:1.18-alpine AS base
 
-RUN apk add build-base
+RUN apk add --no-cache build-base ccls py3-lsp-server
+
+## Build
+FROM base AS build
 
 WORKDIR /app
 
@@ -17,7 +20,7 @@ FROM build AS dev
 
 WORKDIR /app
 
-RUN apk add --no-cache make ccls
+RUN apk add --no-cache make
 
 RUN go install github.com/cespare/reflex@latest
 
@@ -26,11 +29,9 @@ CMD ["make watch"]
 
 
 ## Prod
-FROM alpine:latest AS prod
+FROM base AS prod
 
 WORKDIR /
-
-RUN apk add --no-cache build-base ccls
 
 COPY --from=build /app/server /app/entry.sh /app/.env  /
