+++ go.mod
@@ -6,11 +6,7 @@ require (
 	github.com/fatih/color v1.13.0
 	github.com/joho/godotenv v1.4.0
 	github.com/labstack/echo/v4 v4.9.0
-)
-
-require github.com/gorilla/websocket v1.5.0
-
-require (
+	github.com/gorilla/websocket v1.5.0
 	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
 	github.com/google/uuid v1.3.0
 	github.com/labstack/gommon v0.4.0 // indirect
